+++ src/components/Recommend/Detective/ChatFromBookky.js
@@ -1,8 +1,16 @@
 import styled from "styled-components";
 
-const ChatFromBookky = ({ message }) => {
+const ChatFromBookky = ({ message, hideImg }) => {
   return (
     <ChatFromBookkyContainer>
+      {!hideImg ? (
+        <img
+          src={require("../../../assets/icons/sideNav/welcome.png")}
+          alt="bookky profile"
+        />
+      ) : (
+        <div className="imgArea" />
+      )}
       <div className="message">{message}</div>
     </ChatFromBookkyContainer>
   );
@@ -19,5 +27,16 @@ const ChatFromBookkyContainer = styled.div`
     width: fit-content;
     background-color: var(--bright-base-bg-color);
   }
+
+  img {
+    margin: 0 5px;
+    width: 30px;
+    height: 30px;
+  }
+
+  .imgArea {
+    margin: 0 20px;
+  }
 `;
+
 export default ChatFromBookky;
