+++ src/components/Recommend/Detective/ChatSubmit.js
@@ -1,0 +1,17 @@
+const ChatSubmit = ({ message, hideImg }) => {
+  return (
+    <ChatFromBookkyContainer>
+      {!hideImg ? (
+        <img
+          src={require("../../assets/icons/sideNav/welcome.png")}
+          alt="bookky profile"
+        />
+      ) : (
+        <div className="imgArea" />
+      )}
+      <div className="message">{message}</div>
+    </ChatFromBookkyContainer>
+  );
+};
+
+export default ChatSubmit;
