+++ src/routes/Recommend/Detective.js
@@ -2,8 +2,9 @@ import { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
 import styled from "styled-components";
 import PageHeader from "../../components/PageHeader";
-import ChatFromBookky from "../../components/Recommend/ChatFromBookky";
-import ChatFromUser from "../../components/Recommend/ChatFromUser";
+import ChatFromBookky from "../../components/Recommend/Detective/ChatFromBookky";
+import ChatFromUser from "../../components/Recommend/Detective/ChatFromUser";
+import SelectionBox from "../../components/Recommend/Detective/SelectionBox";
 import Typing from "../../components/Recommend/Typing";
 
 // 추천받개 - 명탐정 북키
@@ -13,42 +14,24 @@ function Detective() {
 
   // User의 대답 여부
   const [answered, setAnswerd] = useState(false);
+  const [onSubmit, setSubmit] = useState(false);
 
   // 채팅 메시지 목록
   const [ChatArray, setChatArray] = useState([
     { mid: "0", who: "bookky", message: "내 이름은 북키, 탐정이죠 🕵🏻‍♀️" },
-    { mid: "1", who: "bookky", message: "네가 찾고 있는 도서를 맞춰볼게 ❗" },
-    { mid: "2", who: "bookky", message: "자바스크립트와 관련한 책이야 ❓" },
+    {
+      mid: "1",
+      who: "bookky",
+      message: "질문에 대답해줘, 도서를 추천해줄게 ❗",
+      hideImg: true,
+    },
+    {
+      mid: "2",
+      who: "bookky",
+      message: "자바스크립트와 관련한 책이야 ❓",
+      hideImg: true,
+    },
   ]);
-
-  // selectAnswer() : 유저의 답변 선택
-  const selectAnswer = (message) => {
-    // 마지막이 유저의 채팅이면, 유저 답변 불가능
-    if (ChatArray[ChatArray.length - 1].who !== "user") {
-      setChatArray([
-        ...ChatArray,
-        { mid: ChatArray.length + 1, who: "user", message: message },
-      ]);
-      setAnswerd(true);
-    }
-  };
-
-  // createBookkyQuestion() : 북키의 질문 생성 (추후 서버로부터 받아오게끔)
-  const createBookkyQuestion = () => {
-    if (answered === true) {
-      setChatArray([
-        ...ChatArray,
-        {
-          mid: ChatArray.length + 1,
-          who: "bookky",
-          message: "파이썬에 관련한 책이야 ❓",
-        },
-      ]);
-      setAnswerd(false);
-    }
-  };
-
-  useEffect(createBookkyQuestion, [answered]);
 
   // 명탐정 북키 View
   return (
@@ -63,55 +46,32 @@ function Detective() {
           <div className="chat">
             {ChatArray.map((el) => {
               if (el.who === "bookky") {
-                return <ChatFromBookky key={el.mid} message={el.message} />;
+                return (
+                  <ChatFromBookky
+                    key={el.mid}
+                    message={el.message}
+                    hideImg={el.hideImg}
+                  />
+                );
               } else if (el.who === "user") {
                 return <ChatFromUser key={el.mid} message={el.message} />;
+              } else if (el.who === "submit") {
               }
             })}
           </div>
         </ChatArea>
         <SelectArea>
-          <SelectionBox>
-            <h1 className="select-header">답변을 선택하세요</h1>
-            <div className="button-area">
-              <button
-                className="answer"
-                onClick={() => selectAnswer("맞아, 확실해 !")}
-              >
-                1️⃣ 맞아, 확실해 !
-              </button>
-              <button
-                className="answer"
-                onClick={() => selectAnswer("전혀 아니야.")}
-              >
-                2️⃣ 전혀 아니야.
-              </button>
-              <button
-                className="answer"
-                onClick={() => selectAnswer("그런 것 같아.")}
-              >
-                3️⃣ 그런 것 같아.
-              </button>
-              <button
-                className="answer"
-                onClick={() => selectAnswer("아닌 것 같아.")}
-              >
-                4️⃣ 아닌 것 같아.
-              </button>
-              <button
-                className="answer"
-                onClick={() => selectAnswer("잘 모르겠는걸 ?")}
-              >
-                5️⃣ 잘 모르겠는걸 ?
-              </button>
-              <button
-                className="submit"
-                onClick={() => selectAnswer("결과를 보여줘 ❗")}
-              >
-                결과를 보여줘 ❗
-              </button>
-            </div>
-          </SelectionBox>
+          {onSubmit ? (
+            <></>
+          ) : (
+            <SelectionBox
+              ChatArray={ChatArray}
+              setAnswerd={setAnswerd}
+              setChatArray={setChatArray}
+              answered={answered}
+              setSubmit={setSubmit}
+            />
+          )}
         </SelectArea>
       </ContentsArea>
     </DetectiveContainer>
@@ -151,7 +111,7 @@ const ChatArea = styled.div`
   background-color: #d7e2ff;
   border-radius: 5px;
   margin: 0 30px;
-  padding: 0 15px;
+  padding: 0 8px;
 
   .name {
     margin: 10px;
@@ -171,51 +131,6 @@ const SelectArea = styled.div`
   min-width: 400px;
   display: flex;
   align-items: center;
-`;
-
-const SelectionBox = styled.div`
-  position: relative;
-  display: grid;
-  grid-template-rows: 20fr 80fr;
-  width: 100%;
-  height: 500px;
-  margin: 0 30px;
-
-  ::before {
-    border-radius: 5px;
-    content: "";
-    position: absolute;
-    z-index: -1;
-    width: 100%;
-    height: 100%;
-    opacity: 80%;
-    background-color: #ffffff;
-    box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
-  }
-
-  .select-header {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 1.5em;
-    font-weight: bold;
-  }
-
-  .button-area {
-    position: relative;
-    display: grid;
-    grid-template-rows: repeat(10fr, 6);
-
-    button:hover {
-      background-color: rgba(110, 149, 255, 0.25);
-    }
-  }
-
-  .submit {
-    font-size: 1.1em;
-    font-weight: bold;
-    color: #6e95ff;
-  }
 `;
 
 export default Detective;
